Finish InsertOrUpdatePerson
